feat: add evaluation module, architecture docs, and improvements
- Add evaluation module for tool call accuracy testing
  - scripts/test.py: CLI for evaluation with DDP support
  - scripts/run_test.sh: Bash wrapper script
  - trm_llm/evaluation/: Core evaluation functions
- Add architecture documentation (docs/architecture.md)
- Add sample data files (tools, system prompt, special tokens)
- Improve tokenizer, inference, and training modules
- Add configs directory

+++ trm_llm/models/reasoning_module.py
@@ -155,7 +155,8 @@ class RecursiveReasoningModule(nn.Module):
 
         # Initialize z if not provided
         if z_current is None:
-            z = torch.zeros(batch_size, self.reasoning_dim, device=x_encoded.device)
+            # Use same dtype as input to support FP16/DeepSpeed
+            z = torch.zeros(batch_size, self.reasoning_dim, device=x_encoded.device, dtype=x_encoded.dtype)
         else:
             z = z_current
 
@@ -223,7 +224,8 @@ class RecursiveReasoningWithSequence(nn.Module):
 
         # Initialize z
         if z_current is None:
-            z = torch.zeros(batch_size, seq_len, self.reasoning_dim, device=x_encoded.device)
+            # Use same dtype as input to support FP16/DeepSpeed
+            z = torch.zeros(batch_size, seq_len, self.reasoning_dim, device=x_encoded.device, dtype=x_encoded.dtype)
         else:
             z = z_current
